Fix: fixed refactor errors in devnetvm
